Split the single controller into 2 separate controllers
